Created todo and quickstart from readme and notes, edited index
I edited the index to reflect a more current state of the site, and what we think/need about it. From README and Notes (both moved to /misc)
I created the quickstart.txt and todo.txt documents. Quickstart includes information about how to get & publish the site with git, basic asciidoc
to get people started editing, some pointers on our site conventions and a primer on css/style. Todo will house a list of pages like to site structure, \
but with some notes on the content and formatting. This is to store ideas on what I need to do and/or what others should contribute for the page and
serve as a loose overview of content once its written. In this commit todo has no new content. Check out quickstart.txt though.