Wrote download page, edited homepage and made some adjustments (still) to todo and quickstart
Primary changes were to download.txt which is scratch written to explain the code status and provide to the
proper links to newbies-tutorial and the appropriate code sources. Tried to maintain a proactive tone.

Tweaked the homepage to be a bit more like the current, made some edits to existing content.

Updated todo to reflect changes in download.txt