Cinelerra woes -- as published 2011
Remark by the comitter(Ichthyo, 2025):
At that time I started to collect notes about the history of Lumiera;
abd somehow I must have found an earlier copy of the Moin-Moin markup
of the same text and transformed that into similarly formatted Asciidoc,
adding a historical note in May 2011.

Since the text in this version is more like a rough draft and contains
several TODO marks, it is likely closer to the version initially
published in summer 2007, which at that time helped to set off
the "Cinelerra-3" initiative. Thus I consider it more adequate
to publish this version here, instead of the more polished
one from the HTML capture in 2008