Kill the Attic
the attic used to hold the leftovers of the old website, and
the lefovers of the "ASCIIDOC day" -- mostly contents from
Cehteh's PiPaWiki, from the early days of the project.

In the last years, I've successively moved over and integrated
most substantial content. Some pages might be of historic interest
though; especially the collection of links

But since everything is preserved in the Git history, I'm removing
these pages from the hidden online presence now.