well -- I meant it rather the other way round
The rvalue refs and variadic templates were the big improvements,
closure turned out to be nice here and there. Probably closures
will start playing a more prominent role in code we write from
now on

And, second correction, I rather wanted to express that
much of this work was boring, tedious and generally not gratifying
It just had to be done.