comment out mention of uWiki in the news archive
explanation: in hindsight this does not look relevant anymore,
since this endeavour stalled soon thereafter, and never got to
the point to host the Lumiera website. Instead, we just stuck
to that nifty solution based on ASCIIDoc and Git, which,
incidentally, was also set up by Christian and was kind of
a first proof-of-concept towards uWiki.