attempt to make the meaning of gestures more pronounced
try to describe it in a more high-level way, with less 'widgets'
and similar technical details

Moreover, I'd prefer not to add the mention of plugins and versions here.
The reason is: while the architecture of Lumiera is indeed structured
such as to allow eventually the addition of commands, and new forms of
interaction through extensions / plug-ins / rules, for the time being
the commands and possible interactions are hard wired and not meant
to be extended. This would just add yet another dimension of flexibility
and is way beyond we can tackle with our limited resources right now.