the new UI-coordinates are topological, not related to mouse position
In fact they complement the existing system of pixel coordinates and
direct widget references, as used by the widget toolkit. Thus the
purpose of the coordinates is not so much related to gestures.
Rather, the goal is to support navigation on a higher, more conceptual
level. And also to express rules of placement / allocation for
view components