Review latest addition on Dependency Injection
Basically this paragraph seems just fine.
However, this topic is not so much currently in debate,
as it was during the last decade. Thus I've changed that
to past.

Moreover, I have moved the reference into a footnote.
And, indeed, this is not just "an interesting article",
this is *the* seminal article which startet the whole debate.

Incidentally, this very article also started this somewhat
unfortunate tradition of using "Inversion of Control" and
"Dependency Injection" within the same context almost as
synonyms. However, I am far from blaming Martin Fowler
for that. Because, if you read carefully, he distinguishes
quite well between the *Principle* (Inversion of Control)
and the *practical solution* for component wiring
(=Dependency Injection).