Switch to HTTPS: also adjust protocol for the ASCIIDOC generated links
the notorious problem with webpage generation: some links need to include
the website hostname and protocol. We use an attribute definition in the
global page.conf, which defines the overall page layout and properties.
Thus the ASCIIDOC page sources can just use {l} or {ldoc} or {img} to
include the respective parts of the URL.

Note: switching to HTTPS also incurs a nasty twist for those people
running a local web server on the local PC for development. Usually you
do not want to bother just run that via HTTP protocol.

This can be accomplished with the help of an ascidoc.conf file on user level

just define (in your home directory)
~/.asciidoc/asciidoc.conf

with the content:

[attributes]
localWebHost=http://localhost:8888

(or whatever you need to get at your local webserver)