Improve build_website.sh
Remark from the committer(Ichthyo, 2025):
This is the start of a series of reworks, which Christian started in 2018.
The result was considered "experimental" and left on website-staging ever since.

As part of the current release and clean-up activities I am now reviewing
and integrating these changes -- however I decided not to accept new
functionality into build_website.sh, like tree cleaning and link checking,
because doing so would stretch this nifty little script beyond its purpose.

Generally speaking, we want to avoid having a ''content management system''