add some xargs magic for more robustness
!!!
Remark from the reviewer(Ichthyo, 2025):
This changeset is borderline to unacceptable:
It introduces a lot of nifty trickery, which relies extensively on the
behavoiur of xargs; furthermore the code is cohesive to such a degree
that I am not able to determine its correctness just from reading it.

However, to my understanding the primary purpose of this change is
to handle file names with embedded spaces properly -- and I don't have
the time and capacity to implement a cleaner solution at the moment.

Thus I accept the "status quo" for now -- under the proviso that we'll
have to re-write our website generator (including Menuegen!) eventually,
in a more approachable and clearly structured way.