Rework and update the project-about and FAQ pages
This text has been added in the early stages of the Lumiera project
and remained largely unchanged ever since. While all the infromation
is basically correct, at various places the focus and scope has
changed. Notably it is not necessary anymore to draw the distinction
to Cinelerra in such a detailed way. On the other hand, several points
were lacking clarity (partially even avoided some clear statements
due to political and contextual considerations).