UPGRADE-TRIXIE: port Menugen to Python-3
Basically this script was written in a ''point-and-shot'' manner a long time ago,
and fulfils its purpose ever since. At that time (2011), python(2) was still the default
on all stable Linux distributions, and the Python-3 effort was still rather new.
Only after v3.5 (2014), Python-3 can be considered the de-facto standard.

Porting the menugen.py turns out to be straight-forward
 * print function
 * fix some imports (`ifilter` ⟶ `filter` built-in)
 * use modernised Exception handling syntax
 * open files with `open()` instead of `file()`
 * check just for `isinstance(x, str)` since we only ever process Unicode
 * in `class ChildDiscoveryRedirected` have to apply one level of iterators,
   since most functional processing in Python-3 now yields iterators, while
   we only need a (flat) iterator over possible paths to visit as result

On that occasion, I tried to understand my own code again and added some
clarification and better naming at places.