Research:Demo-Skeleton: provide a GTK-Application as framework
The topic is to "output video from an desktop application" --
so seemingly we need a ''desktop application'', with GTK being the obvious choice.

This application however will just be used as a ''launchpad'' for the demo code;
- it can be used as access point to connect to some graphics context
- it should provide lifecycle events to start and stop the demo
- it should invoke a frame handling callback

And since the point of this demo is ''not'' do engage in multitasking,
everything should be arranged to run in the GUI-thread.