Research:Animation: much better dither
The first shot was much to elaborate; for this purpose we
do not even need a noise source, it suffices just to mix the bits.

On the other hand, the dither was injected in a clumsy way, so that
it mostly influenced the Luma, leading to quite noticeable noise.
Actually, we really *need* dither only at a single point, which is
when we compute the exponential attenuation. And at that point,
it is easy to inject some random bits per colour channel.