Research:Gtk/Framework: make callback logic more robust
On several instances it could be observed that the periodic callback
was invoked from the GTK-event-loop ''after'' the shutdown hook was invoked.

So seemingly the overall implementation between GLib and GTK is not consistent;
logically, you'd expect a periodic callback to be discontinued before the shutdown
hook is invoked, which seemingly is not the case, events can be processed afterwards.

Thus we have to set up our own guard logic,
which can be achieved by using the callback functors as guards;
a C++ function can be empty and this can be bool-checked.