Research:EGL: get configuration to work
basically I misunderstood one configuration attribute:
EGL_NATIVE_RENDERABLE means in fact that the underlying native display system
is also able to render into the same surface. We rather do *not* want that
(i.e. X11 rendering)

Verified with some query code that we got a sensible config.
Speed is qualitatively the same as with GLX
(and much faster than with SDL, where we got only an X11 impl)