Research:Git: a script to maintain compatibility branches
This is a bash script with some git magic, to manage a
compatibility-branch per user (user name).
By convention, these branches are named "compat/$NAME"

The idea is frequently to re-base such a branch on top of the current branch.

./compat    : rebases on top of current branch
./compat -r : drops back to base branch

Any uncommited changes will be carried along (with the help of git stash)

Usage example

./compat
./compat -r