Research:EGL+GLEW: activate check for successfully init libGLEW
Lib GLEW maintains dynamic bindings for various OpenGL extensions,
based on the current system configuration. It allows to call modern
OpenGL functions directly, without having to check for and load extensions.


Lib GLEW must be initialised /after/ creating the OpenGL context.

On my system this works correct, but the glewInit() call returns
an error code 4, which is translated as "unknown error".

Hopefully this check works correct on a more up-to-date graphics card...