Research:License: decision to publish the demo code under GPL 2+
Initially, our intention was to publish this tutorial and demo code
under a ''permissive license'' -- to allow for mostly unlimited use.

However, during the research it became clear that this whole effort
would never have been possible without the guidance by existing and
reality-proven code bases -- which are published under GPL.


It would be dishonest to imply that this code was created by us alone,
stating with a clean slate, and just by reading the documentation.