Research:IMPROVE Move yuv420 & YV12 to single method and clean-up method
TODO: check to remove comments.
TODO: look over and check. Just flung in the ideas from IRC
and it worked on first compile! So there might be some things
requiring attention.
The idea behind this modification was to set the UV component
offsets outside the main loop depending on either UV12 oor YUV420
and simply increment in offset the U and V components.
Code as is clearly expresses the intention. So we could remove