Clarify License: this code is GPL v2+
To clarify: initially, we intended to publish the demo code
using a permissive license (MIT); however, during the actual research
it became clear that this effort would never have been possible without
the guidance by actual real-world proven code available as Free Software.

Pretending that this code is our invention would be deeply dishonest.

Thus we switched to the GPL already during the research and prior to
the official publication, which happened through the talk at the
FrOSCon conference on 2025-08-16 10:00+2

This changeset merely fixes a leftover mention of the MIT license,
while the file LICENSE already contained a copy of the GPL and all
actual code files were also marked as GPL v2+ at time of publication.