Website/Setup: remove symlink -- handled manually from now on
To simplify the setup of the Lumiera website, the two symlinks
pointing to the documentation section and to the DEB depot were
added to the Git-repository.

However, this causes problems when several website instances
shall be used on the same machine. Thus the symlink is now
removed from the Git-repo and must be setup manually,
so to gain additional flexibility.

See __NOTE_SETUP__ for further information