Webserver-doc: relocate into the infrastructure section
Together with the improvements to the website build-script,
in 2018 Christian also created some pages to document details
of the server and mailinglist setup.

Hereby I relocate these documents into the technical documentation,
where further aspects of our infrastructure are explained.

Note: the first contribution to the build_website.sh dates back to 2009:
2a09bca8d