Link-farm: peliminary manual setup
Since a very long time, we intend automatically to build a directory of keyword cross-links,
to simplify the task of cross-linking to other pages for the webpage author.

See https://Lumiera.org/documentation/devel/rfc/WebsiteSupportMarkup.html

The envisioned automatic solution would require a processing similar
to what is done in the menue generator (`menugen.py`); any usage of a
link to a keyword or index entry in some page would be detected and
result in generation of a suitable link-forwarding page.

Since this task did not make any progress (and presumably won't for the foreseeable future),
I now start to setup a preliminary link-farm manually, because the inability
to use stable cross-links is a major impediment when authoring web pages.
Such a prototypical manual setup might also serve as point-of-reference when
the auto-generation of such pages will be implemented eventually. When this happens,
all manually generated forwarding pages should be removed from Git.