Bait: Food for the trolls!
Improve the bait:
Make the bait linsk visible, lets see if that attracts more bots

Mark the bait-links as disallowed for "robots".
The target for these links does not exist, and we block IP ranges
which produce an unusual high density of 404.

This combination turned out to be extremely successful,
our traffic had been dangerously high and is now reduced below 1/100