Old texts from 2008 -- prepare for conversion
While looking for other stuff related to the Lumiera Website,
I discovered this trove as stale branch in a Git repository.

These texts are HTML captures from Cehteh's wiki at pipapo.org in 2008;
at that time it was already clear that we are moving into the direction
of a Free Software project independent from Cinelerra, and that a new
Website will be built.

The »Cinelerra woes« text was already published since 2011, yet especially
the project proposals seem relevant, since they already describe pretty much
all of the fundamental ideas for Lumiera -- which underpins that the vision
for that new project was formed during a short period between 2007 and 2008
and not changed much ever since.

It seems desirable to integrate those texts into the page with history notes;
thus I'll now start to rewrite these documents into Asciidoc markup