Lumiera History: add page for resources and background information
Since the ''History Page'' as such should be written in the form of a narration,
»the Lumiera story«, to make it a worthwhile read — it becomes evident that a lot
of quite essential detail information needs to be omitted, yet somehow
kept around for later referral.

Thus I intend to use ''footnotes'' as attachment point, from there linking
into appropriate sections of a »Resources« page.