Lumiera History: investigate and document the beginnings
Note: This is an ongoing effort, I'll continue to expand this writing
gradually, and add more background information. The old archives of
the Cinelerra Mailinglist are still online, so that I was able to
incorporate a lot of past circumstances. I have ensured that for
each of these linked old mail archive pages, there is a copy
stored into Archive.org

Furthermore, a massive amount of information has been added
to the »Resources« page, to add some necessary background texture.

- a complete time table of the events from the first year,
  until Lumiera became an independent project

- a comprehensive overview of development activity up to now (2025),
  including participants and relevant "stepstones" in the Git history