Essay: Complexity and Flexibility
The examination of several disputes and controversial topics,
as encountered in the history of Lumiera, inspired me to recall
the insight gained into the nature of software systems, and
especially the topic of essential complexity -- a fundamental
limitation inherent in any larger system, which time and again
shatters our hopes for grandiose technological achievement.

I have written an extended essay to explore this relationship
and to set a proper stage for any further discussion related to
questions of architecture, flexibility and coping with complexity.